% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_res_calc.R
\name{aggregated_res_calc}
\alias{aggregated_res_calc}
\title{Aggregated resistance occurrence calculation}
\usage{
aggregated_res_calc(data, material_summary = "single")
}
\arguments{
\item{data}{Input data}

\item{material_summary}{Default: "single", Use "all" to summarise across material groups, use "food" for summarising for material groups, ignoring animal species}
}
\description{
Function for calculating aggregated resistance occurrences.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
