% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_groups.R
\docType{data}
\name{am_groups}
\alias{am_groups}
\title{am_groups}
\format{
A data frame with 93 rows and 3 variables:
\describe{
  \item{\code{analyttkode_sens}}{character COLUMN_DESCRIPTION}
  \item{\code{substans}}{character COLUMN_DESCRIPTION}
  \item{\code{analyttkode_gruppe}}{character COLUMN_DESCRIPTION}
}
}
\usage{
am_groups
}
\description{
Data frame holding the antimicrobial substance groups
}
\keyword{datasets}
