% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mdr.R
\name{calc_mdr}
\alias{calc_mdr}
\title{Calculate MDR}
\usage{
calc_mdr(data, substance_dist = FALSE, food = FALSE)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{substance_dist}{TRUE/FALSE, default FALSE, Which level of reporting, set to TRUE for percent distribution of each antimicrobial group per MDR category}

\item{food}{Set to TRUE if calculating MDR for food items}
}
\description{
Calculate multi-drug resistance, i.e. how many classes of antibiotics there are resistance against
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
