% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mdr.R
\name{calculate_mdr}
\alias{calculate_mdr}
\title{Calculate MDR}
\usage{
calculate_mdr(data, type = "general", keep_all = NULL, group = "animal")
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{type}{Which level of reporting, leave empty for overall MDR, and "am_groups" for percent distribution of each antimicrobial group per MDR category}

\item{keep_all}{Were all columns kept in the filter_nv_data function? If so, set this to TRUE}

\item{group}{Either "animal" or "food", specify food if calculating MDR for food items}
}
\description{
Calculate multi-drug resistance, i.e. how many classes of antibiotics there are resistance against
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
