% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_res_occurrence.R
\name{calculate_res_occurrence}
\alias{calculate_res_occurrence}
\title{Calculate resistance occurrences}
\usage{
calculate_res_occurrence(
  data,
  material_summary = "single",
  species_summary = "single"
)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{material_summary}{Use "all" for summarising across material groups, use "single" for summarising for each separate material}

\item{species_summary}{Use "all" for summarising across all species groups, use "single" for summarising for each separate species}
}
\description{
Calculate resistance occurrence and confidence intervals on a substance level.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
