% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_res_occurrence_salmonella.R
\name{calculate_res_occurrence_salmonella}
\alias{calculate_res_occurrence_salmonella}
\title{Calculate resistance occurrences for Salmonella}
\usage{
calculate_res_occurrence_salmonella(data, species_summary = "single")
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{species_summary}{Use "all" for summarising across all species groups, use "single" for summarising for each separate species}
}
\description{
Calculate resistance occurrence and confidence intervals on a substance level for Salmonella.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
