% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicated_animal_samples.R
\name{check_duplicated_animal_samples}
\alias{check_duplicated_animal_samples}
\title{Check for duplicates in animal NORM-VET data}
\usage{
check_duplicated_animal_samples(data, year, animal_species)
}
\arguments{
\item{data}{Input data}

\item{animal_species}{The animal species of interest, can be a vector}
}
\description{
Function for checking for duplicates in the NORM-VET animal data, based on producer
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
