% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_efsa_data.R
\name{check_efsa_data}
\alias{check_efsa_data}
\title{Check EFSA data for errors}
\usage{
check_efsa_data(
  data,
  empty_check_cols,
  check_operator = TRUE,
  check_dates = TRUE
)
}
\arguments{
\item{data}{Input data}

\item{empty_check_cols}{Vector with column names to check for empty values}

\item{check_operator}{If TRUE, will check for incorrect operators in the data}

\item{check_dates}{If TRUE, will check for incorrect dates in the data}
}
\description{
Function for checking prepared EFSA data for submission to the EFSA system.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
