% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_efsa_dates.R
\name{check_efsa_dates}
\alias{check_efsa_dates}
\title{Check for date errors in EFSA data}
\usage{
check_efsa_dates(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Function for checking EFSA data for incorrect dates
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
