% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_efsa_empty_values.R
\name{check_efsa_empty_values}
\alias{check_efsa_empty_values}
\title{Check for missing values in EFSA data}
\usage{
check_efsa_empty_values(data, empty_check_cols)
}
\arguments{
\item{data}{Input data}

\item{empty_check_cols}{A vector of column names of interest}
}
\description{
Function for identifying missing values in EFSA data columns
}
\author{
Håkon Kaspersen, \email{hakon.pedersen.kaspersen@vetinst.no}
}
