% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_efsa_operator.R
\name{check_efsa_operator}
\alias{check_efsa_operator}
\title{Check for operator errors in EFSA data}
\usage{
check_efsa_operator(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Function for checking EFSA data for incorrect operators
}
\author{
Håkon Kaspersen, \email{hakon.pedersen.kaspersen@vetinst.no}
}
