% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mic_and_phenotype.R
\name{create_mic_and_phenotype}
\alias{create_mic_and_phenotype}
\title{Clean up data from NORM-VET database}
\usage{
create_mic_and_phenotype(data, retain_plate2 = FALSE)
}
\arguments{
\item{data}{The data frame from fetch_nv_data}

\item{retain_plate2}{Set to TRUE if you do not wish to filter out plate2 results (020185)}
}
\description{
Run this function to generate corrected MIC values and phenotype definitions, as well as removing redundant columns.
}
\author{
Håkon Kaspersen \email{hakon.kaspersen@vetinst.no}
}
