% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_micdist.R
\name{create_micdist}
\alias{create_micdist}
\title{Create MIC distribution tables}
\usage{
create_micdist(
  data,
  bacteria_category = NULL,
  material_group = NULL,
  bacteria_group = NULL,
  species_group = NULL,
  reporting_year = NULL
)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{bacteria_category}{Bacterial category to filter on}

\item{material_group}{Material group to filter on, leave blank for summarising across material groups}

\item{bacteria_group}{Bacteria group to filter on}

\item{species_group}{Animal species to filter on, leave blank to include all species}

\item{reporting_year}{Report year to filter on}
}
\description{
Script for creating MIC distribution tables from resistance data.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
