% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_micdist_salmonella.R
\name{create_micdist_salmonella}
\alias{create_micdist_salmonella}
\title{Create MIC distribution table for Salmonella}
\usage{
create_micdist_salmonella(
  data,
  bacteria_category = NULL,
  species_group = NULL,
  reporting_year = NULL,
  material = NULL
)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{bacteria_category}{Bacterial category to filter on}

\item{species_group}{Animal species to filter on, leave blank to include all species}

\item{reporting_year}{Report year to filter on}

\item{material}{Either "Mat" or "Dyr"}
}
\description{
Script for creating MIC distribution tables from resistance data for Salmonella
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
