% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_data.R
\docType{data}
\name{cutoff_data}
\alias{cutoff_data}
\title{cutoff_data}
\format{
A data frame with 503 rows and 7 variables:
\describe{
  \item{\code{cut_off_gruppe}}{character COLUMN_DESCRIPTION}
  \item{\code{analyttkode_sens}}{character COLUMN_DESCRIPTION}
  \item{\code{cut_off}}{numeric COLUMN_DESCRIPTION}
  \item{\code{dato}}{character COLUMN_DESCRIPTION}
  \item{\code{kilde}}{character COLUMN_DESCRIPTION}
  \item{\code{mo}}{character COLUMN_DESCRIPTION}
  \item{\code{substans}}{character COLUMN_DESCRIPTION}
}
}
\usage{
cutoff_data
}
\description{
Data frame holding the antimicrobial cutoff values
}
\keyword{datasets}
