% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsa_codes.R
\docType{data}
\name{efsa_codes}
\alias{efsa_codes}
\title{efsa_codes}
\format{
A list with 11 data frames
}
\usage{
efsa_codes
}
\description{
A list of data frames holding the various EFSA codes
}
\keyword{datasets}
