% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsa_mics_and_ranges.R
\name{efsa_mics_and_ranges}
\alias{efsa_mics_and_ranges}
\title{Correct to EFSA MICs and ranges}
\usage{
efsa_mics_and_ranges(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Function for correcting MIC-values and ranges to EFSA standards
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
