% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nv_data.R
\name{fetch_nv_data}
\alias{fetch_nv_data}
\title{Connect to NORM-VET database and fetch data}
\usage{
fetch_nv_data(server, database, user)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}

\item{pw_method}{Specifies how the user supplies the database password, either "prompt" for an actual promt, or "local" to specify a text file holding the password}

\item{pw_file_path}{Path to text file holding the password, must be supplied when using "pw_method = local"}
}
\description{
This function connects to the NORM-VET database and merges data in the correct format. Here, information on samples and results are imported, but not MIC values. Users will have to supply their password to the database manually when running the function.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
