% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nv_data_aggregated.R
\name{filter_nv_data_aggregated}
\alias{filter_nv_data_aggregated}
\title{Filter data based on user input}
\usage{
filter_nv_data_aggregated(
  data,
  bacteria_category = NULL,
  material_group = NULL,
  bacteria_group = NULL,
  species_group = NULL,
  reporting_year = NULL
)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{bacteria_category}{Bacterial category to filter on}

\item{material_group}{Material group to filter on, leave blank for summarising across material groups}

\item{bacteria_group}{Bacteria group to filter on}

\item{species_group}{Animal species to filter on, leave blank to include all species}

\item{reporting_year}{Report year to filter on}
}
\description{
Function for filtering data for aggregated analysis and throwing error if filtered data is empty.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
