% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_data.R
\docType{data}
\name{gene_data}
\alias{gene_data}
\title{gene_data}
\format{
A data frame with 489 rows and 2 variables:
\describe{
  \item{\code{analyttkode}}{character analyte code from the database}
  \item{\code{gene}}{character gene name from the database}
}
}
\usage{
gene_data
}
\description{
Data frame holding the gene codes and names
}
\keyword{datasets}
