% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_empty_cols.R
\docType{data}
\name{gene_empty_cols}
\alias{gene_empty_cols}
\title{gene_empty_cols}
\format{
A vector with 26 elements
}
\usage{
gene_empty_cols
}
\description{
Data vector holding the column names used to check for empty column values in the EFSA gene data
}
\keyword{datasets}
