% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_efsa_data.R
\name{generate_efsa_data}
\alias{generate_efsa_data}
\title{Generate EFSA data}
\usage{
generate_efsa_data(
  data,
  n_df,
  year,
  category,
  bacteria_group,
  animal_species,
  exclude_combinations = NULL
)
}
\arguments{
\item{data}{Input data}

\item{n_df}{Data frame holding the total number of samples per combination}

\item{year}{Relevant year to filter on (only one allowed)}

\item{category}{The bacterial categories of interest, may be a vector}

\item{bacteria_group}{The bacterial groups of interest, may be a vector}

\item{animal_species}{The animal species of interest, may be a vector}

\item{exclude_combinations}{A data frame with unwanted combinations. The data frame holds three columns: bakterie_kategori, bakterie_gruppe, and art_gruppe}
}
\description{
Function for generating the main EFSA reporting data
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
