% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_efsa_dates.R
\name{generate_efsa_dates}
\alias{generate_efsa_dates}
\title{Generate EFSA-specific dates for reporting data}
\usage{
generate_efsa_dates(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Function for generating EFSA-specific sampling date, isolation date, and analysis date
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
