% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_efsa_neg_data.R
\name{generate_efsa_neg_data}
\alias{generate_efsa_neg_data}
\title{Generate EFSA data for Negative data model}
\usage{
generate_efsa_neg_data(data, year, bacteria_group, animal_species)
}
\arguments{
\item{data}{Input data}

\item{year}{Relevant year to filter on (only one allowed)}

\item{bacteria_group}{The bacterial groups of interest, may be a vector}

\item{animal_species}{The animal species of interest, may be a vector}
}
\description{
Function for generating EFSA data for the negative data model
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
