% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binCI.R
\name{get_binCI}
\alias{get_binCI}
\title{Calculate Confidence Intervals}
\usage{
get_binCI(x, n)
}
\arguments{
\item{x}{The name of the column with the number of positive/resistant samples}

\item{n}{The name of the column with the total number of samples}
}
\description{
This function is used to calculate confidence intervals.
}
\author{
Håkon Kaspersen \email{hakon.kaspersen@vetinst.no}
}
