% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_efsa_codes.R
\name{join_efsa_codes}
\alias{join_efsa_codes}
\title{Join EFSA code data to data frame}
\usage{
join_efsa_codes(data, gene_data = FALSE)
}
\arguments{
\item{data}{Input data}

\item{gene_data}{If TRUE, will merge with gene-related EFSA data}
}
\description{
Function for joining EFSA code data to existing data frame
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
