% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_filtering_values.R
\name{list_filtering_values}
\alias{list_filtering_values}
\title{List filtering values}
\usage{
list_filtering_values(server, database, user)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}
}
\description{
Function for listing filtering values from the database groups.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
