% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nv_tables.R
\name{list_nv_tables}
\alias{list_nv_tables}
\title{List database information}
\usage{
list_nv_tables()
}
\description{
This function returns a data frame with information on each table used in the database.
}
\author{
Håkon Kaspersen \email{hakon.kaspersen@vetinst.no}
}
