% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{Colour palettes used in noRmvet}
\usage{
list_palettes(group = NULL)
}
\arguments{
\item{group}{Which group palette you want to extract, either "species", "food, "am_groups", or "mdr"}
}
\description{
This function holds the colour palettes used in the plot functions.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
