% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr.R
\name{mdr}
\alias{mdr}
\title{Calculate multi-drug resistance distributions}
\usage{
mdr(
  data,
  bacteria_category = NULL,
  material_group = NULL,
  bacteria_group = NULL,
  species_group = NULL,
  reporting_year = NULL,
  substance_dist = FALSE,
  food = FALSE
)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{bacteria_category}{Bacterial category to filter on}

\item{material_group}{Material group to filter on}

\item{bacteria_group}{Bacteria group to filter on}

\item{species_group}{Animal species to filter on}

\item{reporting_year}{Report year to filter on}

\item{substance_dist}{TRUE/FALSE, default FALSE, Which level of reporting, set to TRUE for percent distribution of each antimicrobial group per MDR category}

\item{food}{Set to TRUE if calculating MDR for food items}
}
\description{
Calculate the multi-drug resistance distribution, i.e. how many classes of antibiotics there are resistance against
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
