% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_empty_cols.R
\docType{data}
\name{mic_empty_cols}
\alias{mic_empty_cols}
\title{mic_empty_cols}
\format{
A vector with 30 elements
}
\usage{
mic_empty_cols
}
\description{
Data vector holding the column names used to check for empty column values in the EFSA MIC data
}
\keyword{datasets}
