% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ranges.R
\docType{data}
\name{panel_ranges}
\alias{panel_ranges}
\title{panel_ranges}
\format{
A data frame with 1418 rows and 8 variables:
\describe{
  \item{\code{report_year}}{character COLUMN_DESCRIPTION}
  \item{\code{bakterie_gruppe}}{character COLUMN_DESCRIPTION}
  \item{\code{bakterie_kategori}}{character COLUMN_DESCRIPTION}
  \item{\code{panel}}{character COLUMN_DESCRIPTION}
  \item{\code{plate_def}}{character COLUMN_DESCRIPTION}
  \item{\code{substans}}{character COLUMN_DESCRIPTION}
  \item{\code{range_min}}{numeric COLUMN_DESCRIPTION}
  \item{\code{range_max}}{numeric COLUMN_DESCRIPTION}
}
}
\usage{
panel_ranges
}
\description{
Data frame holding the panels and ranges for all substances for each year and group
}
\keyword{datasets}
