% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdr.R
\name{plot_mdr}
\alias{plot_mdr}
\title{Plot MDR}
\usage{
plot_mdr(data, type = "general")
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{type}{Which level of reporting, use "general" for overall MDR, and "am_groups" for percent distribution of each antimicrobial group per MDR category}
}
\description{
Plot multi-drug resistance, i.e. how many classes of antibiotics there are resistance against
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
