% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrence_over_time.R
\name{plot_occurrence_over_time}
\alias{plot_occurrence_over_time}
\title{Plot occurrence over time lineplot}
\usage{
plot_occurrence_over_time(data, am_group = NULL, species_group = NULL)
}
\arguments{
\item{data}{The data holding the MIC values and sample information}

\item{am_group}{Antimicrobial groups to filter on}

\item{species_group}{Animal species to filter on}
}
\description{
Function for creating line plots with resistance occurrence of each antimicrobial group over time.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
