% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_calc.R
\name{res_calc}
\alias{res_calc}
\title{Resistance occurrence calculation}
\usage{
res_calc(data, material_summary = "single")
}
\arguments{
\item{data}{Input data}

\item{material_summary}{Default: "single", Use "all" to summarise across material groups, use "food" for summarising for material groups, ignoring animal species}
}
\description{
Function for calculating resistance occurrences.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
