% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{species_codes}
\format{
A data frame with 256 rows and two columns:
\describe{
  \item{\code{artkode}}{character The species codes from PJS}
  \item{\code{art}}{character The corresponding species name}
}
}
\usage{
species_codes
}
\description{
Data frame holding the species codes from pjs
}
\keyword{datasets}
