% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_order.R
\docType{data}
\name{substance_order}
\alias{substance_order}
\title{substance_order}
\format{
A vector with 75 elements
}
\usage{
substance_order
}
\description{
Data vector holding substance orders for all substances
}
\keyword{datasets}
