% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_test.R
\name{syn_test}
\alias{syn_test}
\title{Determine synergy}
\usage{
syn_test(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Function for determining synergy between cefotaxime and cefotaxime/clavulanic acid, and ceftazidime and ceftazidime/clavulanic acid
}
\author{
Håkon Kaspersen, \email{hakon.pedersen.kaspersen@vetinst.no}
}
