% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_analyte_sens_group.R
\name{update_analyte_sens_group}
\alias{update_analyte_sens_group}
\title{Update antimicrobial codes and names}
\usage{
update_analyte_sens_group(server, database, user)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}
}
\description{
This function updates the "analytt_sens_group" table in the NORM-VET Database.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
