% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cutoff.R
\name{update_cutoff}
\alias{update_cutoff}
\title{Update cutoff values for antimicrobials}
\usage{
update_cutoff(server, database, user)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}
}
\description{
This function updates the "analytt_sens_cutoff" table in the NORM-VET Database.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
