% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_material_group.R
\name{update_material_group}
\alias{update_material_group}
\title{Update material groups}
\usage{
update_material_group(server, database, user, update = FALSE)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}

\item{update}{Logical, whether or not the data will be sent to the database. Check differences within the data before updating the database by using `update = FALSE`}
}
\description{
This function updates the "materiale_gruppe" table in the NORM-VET Database. The variable "update" is by default set to FALSE, so that the user can check the output data before overwriting the existing data in the database. The output of the function is a list containing three elements; the old data, the new data, and the differences between them.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
