% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_efsa_xml.R
\name{write_efsa_xml}
\alias{write_efsa_xml}
\title{Write EFSA xml file}
\usage{
write_efsa_xml(data, path, root = "dataset", row_tag = "result")
}
\arguments{
\item{data}{Input data}

\item{path}{Output file path and name}

\item{root}{Name of root in the xml}

\item{row_tag}{Name of rows in the xml}
}
\description{
Function for writing the EFSA xml file
}
\author{
Håkon Kaspersen, \email{hakon.pedersen.kaspersen@vetinst.no}
}
