% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_escape.R
\name{xml_escape}
\alias{xml_escape}
\title{Xml escape}
\usage{
xml_escape(x)
}
\arguments{
\item{x}{Input data}
}
\description{
Function for escaping characters in xml output
}
\author{
Håkon Kaspersen, \email{hakon.pedersen.kaspersen@vetinst.no}
}
