% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_samples.R
\name{count_samples}
\alias{count_samples}
\title{Count number of samples}
\usage{
count_samples(data, salmonella = FALSE, year = NULL)
}
\arguments{
\item{data}{Input data}

\item{salmonella}{If TRUE, will return data on Salmonella samples}

\item{year}{Year to filter on}
}
\description{
Function for calculating the number of samples in the database.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
