% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_data.R
\name{create_gene_data}
\alias{create_gene_data}
\title{Fetch gene data from the raw NORM-VET data}
\usage{
create_gene_data(data)
}
\arguments{
\item{data}{The data frame from fetch_nv_data}
}
\description{
Run this function to generate a data frame holding all the gene data in the NORM-VET database
}
\author{
Håkon Kaspersen \email{hakon.kaspersen@vetinst.no}
}
