% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nv_mic_data.R
\name{fetch_nv_mic_data}
\alias{fetch_nv_mic_data}
\title{Connect to NORM-VET database and fetch MIC data}
\usage{
fetch_nv_mic_data(
  server = NULL,
  database = NULL,
  user = NULL,
  password = NULL,
  pw_prompt = TRUE
)
}
\arguments{
\item{server}{Name of the server to connect to}

\item{database}{Name of the database to fetch data from}

\item{user}{Username for the database connection}

\item{password}{Password for the database connection, leave blank if pw_prompt = TRUE}

\item{pw_prompt}{Specifies how the user supplies the database password, TRUE means a prompt will appear, FALSE means the user will supply the password directly}
}
\description{
This function connects to the NORM-VET database and merges data in the correct format. Here, all data, including MIC values are imported. Users will have to supply their password to the database manually when running the function.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
